+++ internal/assets/views/index.html
@@ -188,6 +188,6 @@
     </audio>
     <audio id="backgroundMusic" src="assets/music-slow.mp3" preload="auto" loop></audio>
     <audio id="backgroundBossMusic" src="assets/music.mp3" preload="auto" loop></audio>
-    <script src="main.js" defer></script>
+    <script type="module" src="js/main.js" defer></script>
 </body>
 </html>
