+++ internal/assets/views/js/entities/index.js
@@ -1,0 +1,7 @@
+// Entity Classes Export
+export { Player } from './player.js';
+export { Projectile, InvaderProjectile } from './projectiles.js';
+export { Invader } from './invader.js';
+export { Boss, setBossImage } from './boss.js';
+export { Grid } from './grid.js';
+export { Particle, FlashingText } from './particle.js';
